feat: Update databaseFunctions.ts and timetableDate.ts
This commit updates the code in databaseFunctions.ts and timetableDate.ts files. In databaseFunctions.ts, the import statements for webuntis and prisma client are changed to use double quotes instead of single quotes. Additionally, a new function getTeacherData is added to retrieve teacher data from the database. In timetableDate.ts, the formatUntisTime function is now awaited when formatting lesson start and end times. This ensures that the times are correctly formatted before being used. These changes improve the functionality and maintainability of the code.