Update schema.prisma and DBUpdater.ts
This commit updates the schema.prisma file to fix a typo in the relation field name. The "Class" relation field is changed to "class" to match the naming convention. In DBUpdater.ts, the code is modified to fetch and save data for the "Room" and "Subject" models. The getAllRooms function is added to fetch all rooms from the API and mapToRoomModel function is implemented to transform the data into the required format. Similarly, the getAllSubjects function is used to fetch all subjects and mapToSubjectModel function is implemented to map the data. These changes improve the accuracy and completeness of the database data.